+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This file documents all updates and releases to the ResearchNow Data Ingest pipe
 ## [0.0.5] - Unreleased
 ### Added
 - Edition detail endpoint to the API to allow users to retrieve an individual edition and its component instances
+- travisCI deployment for search API
 
 ## [0.0.4] - 2020-04-16
 ### Added